backend/requirements.txt
PyMuPDF==1.27.1
Pillow==12.1.1
pytesseract==0.3.13
